requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
mysqlclient==2.2.7
numpy==2.2.4
packaging==25.0
pluggy==1.6.0
protobuf==4.21.12
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.7
